+++ LuckyoneUI/Core/Core.lua
@@ -5,7 +5,7 @@ local LDB = Private.Libs.LDB
 local LDBI = Private.Libs.LDBI
 
 -- Lua functions
-local format = format
+local format = string.format
 local ipairs = ipairs
 local next = next
 local pairs = pairs
@@ -32,7 +32,8 @@ local SetCVar = C_CVar.SetCVar
 local _G = _G
 
 -- Blizzard functions
-local Settings_OpenToCategory = _G.Settings.OpenToCategory
+local Settings_OpenToCategory = _G.Settings and _G.Settings.OpenToCategory
+local InterfaceOptionsFrame_OpenToCategory = _G.InterfaceOptionsFrame_OpenToCategory
 local StaticPopup_Show = _G.StaticPopup_Show
 
 -- Global strings
@@ -92,8 +93,13 @@ local LuckyoneLDB = LDB:NewDataObject(Name, {
 			if Private.ElvUI then
 				ElvUI[1]:ToggleOptions()
 				ElvUI[1].Libs.AceConfigDialog:SelectGroup('ElvUI', 'LuckyoneUI')
+				ElvUI[1]:Config_UpdateSize(true)
 			else
-				Settings_OpenToCategory('LuckyoneUI')
+				if Settings_OpenToCategory and Private.SettingsCategoryID then
+					Settings_OpenToCategory(Private.SettingsCategoryID)
+				elseif InterfaceOptionsFrame_OpenToCategory and Private.SettingsFrame then
+					InterfaceOptionsFrame_OpenToCategory(Private.SettingsFrame)
+				end
 			end
 		elseif button == 'RightButton' then
 			if IsShiftKeyDown() then
@@ -120,8 +126,13 @@ function LuckyoneUI_OnAddonCompartmentClick()
 	if Private.ElvUI then
 		ElvUI[1]:ToggleOptions()
 		ElvUI[1].Libs.AceConfigDialog:SelectGroup('ElvUI', 'LuckyoneUI')
+		ElvUI[1]:Config_UpdateSize(true)
 	else
-		Settings_OpenToCategory('LuckyoneUI')
+		if Settings_OpenToCategory and Private.SettingsCategoryID then
+			Settings_OpenToCategory(Private.SettingsCategoryID)
+		elseif InterfaceOptionsFrame_OpenToCategory and Private.SettingsFrame then
+			InterfaceOptionsFrame_OpenToCategory(Private.SettingsFrame)
+		end
 	end
 end
 
@@ -263,8 +274,13 @@ function Private.Addon:Toggles(msg)
 		if Private.ElvUI then
 			ElvUI[1]:ToggleOptions()
 			ElvUI[1].Libs.AceConfigDialog:SelectGroup('ElvUI', 'LuckyoneUI')
+			ElvUI[1]:Config_UpdateSize(true)
 		else
-			Settings_OpenToCategory('LuckyoneUI')
+			if Settings_OpenToCategory and Private.SettingsCategoryID then
+				Settings_OpenToCategory(Private.SettingsCategoryID)
+			elseif InterfaceOptionsFrame_OpenToCategory and Private.SettingsFrame then
+				InterfaceOptionsFrame_OpenToCategory(Private.SettingsFrame)
+			end
 		end
 	elseif msg == 'minimap' then
 		if Private.Addon.db.profile.minimap.hide then
@@ -394,10 +410,11 @@ function Private:HandleToons()
 		['Player-4454-060E364E'] = true, -- [H] Priest
 		['Player-4454-060E361A'] = true, -- [H] Shaman
 	} or Private.isTBC and {
-		-- (6112: Spineshatter)
-		['Player-6112-028A3A6D'] = true, -- [H] Hunter
-		['Player-6112-02A2F754'] = true, -- [H] Priest
-		['Player-6112-02A39E0E'] = true, -- [H] Warlock
+		-- (6412: Spineshatter)
+		['Player-6412-028A3A6D'] = true, -- [H] Hunter
+		['Player-6412-0336641F'] = true, -- [H] Priest
+		['Player-6412-02A39E0E'] = true, -- [H] Warlock
+		['Player-6412-02BBE8AB'] = true, -- [H] Hunter 2
 	} or Private.isClassic and {
 		-- (5827: Living Flame)
 		['Player-5827-0273D732'] = true, -- [A] Hunter
@@ -436,6 +453,12 @@ function Private.Addon:PLAYER_ENTERING_WORLD(_, initLogin, isReload)
 	Private:EasyDelete()
 	Private:HandleToons()
 	Private:PrivacyOverlay()
+
+	if Private.ElvUI then
+		Private:MythicVisibility()
+		Private:DataTextsTweaks()
+	end
+
 	self:LoadCommands()
 
 	if Private.itsLuckyone then
@@ -453,8 +476,13 @@ function Private.Addon:PLAYER_LOGIN()
 	end
 end
 
+function Private.Addon:PLAYER_SPECIALIZATION_CHANGED()
+	Private:DataTextsTweaks()
+end
+
 -- Register events during addon initialization
 function Private.Addon:RegisterEvents()
 	self:RegisterEvent('PLAYER_ENTERING_WORLD')
 	self:RegisterEvent('PLAYER_LOGIN')
+	self:RegisterEvent('PLAYER_SPECIALIZATION_CHANGED')
 end