+++ CHANGELOG.md
@@ -1,4 +1,28 @@
-﻿## Version 3.32 [14-01-2026]
+﻿## Version 3.33 [21-01-2026]
+
+**All WoW Versions:**  
+- ElvUI 14.07+ compatibility & ElvUI Midnight PTR build support  
+
+**Retail:**  
+- Converted some of my custom ElvUI WeakAuras to options in the Developer config tab  
+- New basic text format tags (based on Blizzard restrictions) added  
+- OmniCD profile removed (Cooldown and interrupt tracking is restricted)  
+- Re-running the installer for all layouts and profiles is required  
+- This version is the initial support for Midnight patch 12.0 (WIP!)  
+- Updated Plater profile for Midnight (WIP! I still use ElvUI Nameplates)  
+
+**TBC:**  
+- Fixed energy and mp5 ticks not getting enabled for fresh installs  
+- Ported my custom skins for various AddOn elements into my AddOn  
+- DejaClassicStats  
+- Leatrix Plus  
+- LFG Bulletin Board  
+- Nova Instance Tracker  
+- Nova Spell Rank Checker  
+- Nova World Buffs  
+- WhatsTraining  
+
+## Version 3.32 [14-01-2026]
 
 **All WoW Versions:**  
 - ElvUI 14.06+ compatibility